# Optimize `pop!(s::Set, x, default)`