@jakobnissen 2023-11-04
I'm curious why it is slower for x_negative?
Anyway, your benchmark is the best possible case for the implementation on master - if hashing is more expensive, this PRs implementation will pull ahead.