@petvana 2023-11-04
Yes, the speedup is about 30% here, but should be close to 50% for expensive hashing function. 

The slowdown for `x_negative` (returning default value) is almost negligible and I guess it depends on specific clang optimizations.